Dev/2.2.1 add pipe

+++ src/main/java/io/github/makbn/jthumbnail/core/util/IOUtil.java
@@ -41,6 +41,10 @@ public class IOUtil {
     public static void deleteQuietlyForce(File file) {
         if (file != null) {
             try {
+                if (file.getAbsolutePath().equalsIgnoreCase(System.getProperty("java.io.tmpdir"))) {
+                    log.debug("Can't delete system tmp folder: {}! skipping...", file.getAbsolutePath());
+                    return;
+                }
                 Files.deleteIfExists(file.toPath());
             } catch (IOException e) {
                 log.error(e.getMessage(), e);
